src/promotion/nats-subscriber/requirements.txt
nats-py==2.7.2
aiohttp==3.10.1